backend/requirements.txt
flask
python-dotenv
supabase
torch
transformers
numpy
pandas
fastapi
uvicorn[standard]
pydantic
